Basic logic for turning on LEDs remotely works now.
It is now possible to turn LEDs on remotely by accessing the "WWW-server"
on the device. The device recognizes LEDs to turn on.

Needs some fine tuning. Not always stable. Need to add LED toggling and LED off.
Need to remove debug LED calls to not interfere with normal operation.